+++ src/TestFramework/PhpUnit/Adapter/PhpUnitAdapterFactory.php
@@ -47,12 +47,10 @@ use Infection\TestFramework\PhpUnit\Config\Builder\MutationConfigBuilder;
 use Infection\TestFramework\PhpUnit\Config\Path\PathReplacer;
 use Infection\TestFramework\PhpUnit\Config\XmlConfigurationManipulator;
 use Infection\TestFramework\PhpUnit\Config\XmlConfigurationVersionProvider;
-use Infection\TestFramework\SafeDOMXPath;
 use Infection\TestFramework\VersionParser;
 use function Safe\file_get_contents;
 use SplFileInfo;
 use Symfony\Component\Filesystem\Filesystem;
-use Throwable;
 use Webmozart\Assert\Assert;
 
 /**
@@ -80,7 +78,6 @@ final class PhpUnitAdapterFactory implements TestFrameworkAdapterFactory
         Assert::string($testFrameworkConfigDir, 'Config dir is not allowed to be `null` for the adapter');
 
         $testFrameworkConfigContent = file_get_contents($testFrameworkConfigPath);
-        $requireCoverageMetadata = self::parseRequireCoverageMetadata($testFrameworkConfigContent);
 
         $configManipulator = new XmlConfigurationManipulator(
             new PathReplacer(
@@ -117,7 +114,7 @@ final class PhpUnitAdapterFactory implements TestFrameworkAdapterFactory
                 $executeOnlyCoveringTestCases,
                 $filteredSourceFilesToMutate,
                 $mapSourceClassToTestStrategy,
-                $requireCoverageMetadata,
+                $testFrameworkConfigContent,
             ),
             new VersionParser(),
             new CommandLineBuilder(),
@@ -132,16 +129,5 @@ final class PhpUnitAdapterFactory implements TestFrameworkAdapterFactory
     public static function getExecutableName(): string
     {
         return 'phpunit';
-    }
-
-    private static function parseRequireCoverageMetadata(string $xmlContent): bool
-    {
-        try {
-            $xPath = SafeDOMXPath::fromString($xmlContent, preserveWhiteSpace: false);
-
-            return $xPath->queryAttribute('/phpunit/@requireCoverageMetadata')?->nodeValue === 'true';
-        } catch (Throwable) {
-            return false;
-        }
     }
 }
